Remove state analyzer and join to block analyzer: new name ChainAnalyzer
