+++ pkg/analyzer/block.go
@@ -10,37 +10,47 @@ import (
 	"github.com/cortze/eth-cl-state-analyzer/pkg/db"
 	prom_metrics "github.com/cortze/eth-cl-state-analyzer/pkg/metrics"
 	"github.com/cortze/eth-cl-state-analyzer/pkg/spec"
+	"github.com/cortze/eth-cl-state-analyzer/pkg/spec/metrics"
+	"github.com/sirupsen/logrus"
 
 	"github.com/cortze/eth-cl-state-analyzer/pkg/events"
 	"github.com/pkg/errors"
 )
 
-type BlockAnalyzer struct {
-	ctx       context.Context
-	cancel    context.CancelFunc
+type ChainAnalyzer struct {
+	ctx    context.Context
+	cancel context.CancelFunc
+
+	// Slot Range
 	initSlot  phase0.Slot
 	finalSlot phase0.Slot
 
+	// Channels
+	epochTaskChan       chan *EpochTask
+	valTaskChan         chan *ValTask
 	blockTaskChan       chan *BlockTask
 	transactionTaskChan chan *TransactionTask
 
-	cli      *clientapi.APIClient
-	dbClient *db.PostgresDBService
+	// Connections
+	cli       *clientapi.APIClient
+	eventsObj events.Events
+	dbClient  *db.PostgresDBService
 
-	downloadMode string
+	downloadMode       string
+	validatorWorkerNum int
+	metrics            DBMetrics
+
 	// Control Variables
 	stop              bool
 	downloadFinished  bool
 	processerFinished bool
 	routineClosed     chan struct{}
-	eventsObj         events.Events
 
-	initTime           time.Time
-	enableTransactions bool
-	PromMetrics        *prom_metrics.PrometheusMetrics
+	initTime    time.Time
+	PromMetrics *prom_metrics.PrometheusMetrics
 }
 
-func NewBlockAnalyzer(
+func NewChainAnalyzer(
 	pCtx context.Context,
 	httpCli *clientapi.APIClient,
 	initSlot uint64,
@@ -49,9 +59,9 @@ func NewBlockAnalyzer(
 	workerNum int,
 	dbWorkerNum int,
 	downloadMode string,
-	enableTransactions bool,
-	prometheusPort int) (*BlockAnalyzer, error) {
-	log.Infof("generating new Block Analzyer from slots %d:%d", initSlot, finalSlot)
+	metrics string,
+	prometheusPort int) (*ChainAnalyzer, error) {
+	log.Infof("generating new Block Analyzer from slots %d:%d", initSlot, finalSlot)
 	// gen new ctx from parent
 	ctx, cancel := context.WithCancel(pCtx)
 
@@ -61,6 +71,12 @@ func NewBlockAnalyzer(
 	if downloadMode == "hybrid" || downloadMode == "historical" {
 		log.Debug("slotRanges are:", slotRanges)
 	}
+
+	metricsObj, err := NewMetrics(metrics)
+	if err != nil {
+		return nil, errors.Wrap(err, "unable to read metric.")
+	}
+
 	idbClient, err := db.New(ctx, idbUrl, db.WithWorkers(dbWorkerNum))
 	if err != nil {
 		return nil, errors.Wrap(err, "unable to generate DB Client.")
@@ -71,19 +87,22 @@ func NewBlockAnalyzer(
 	// generate the central exporting service
 	promethMetrics := prom_metrics.NewPrometheusMetrics(ctx, "0.0.0.0", prometheusPort)
 
-	analyzer := &BlockAnalyzer{
+	analyzer := &ChainAnalyzer{
 		ctx:                 ctx,
 		cancel:              cancel,
 		initSlot:            phase0.Slot(initSlot),
 		finalSlot:           phase0.Slot(finalSlot),
+		epochTaskChan:       make(chan *EpochTask, 1),
+		valTaskChan:         make(chan *ValTask, workerNum),
 		blockTaskChan:       make(chan *BlockTask, 1),
 		transactionTaskChan: make(chan *TransactionTask, 1),
+		validatorWorkerNum:  workerNum,
 		cli:                 httpCli,
 		dbClient:            idbClient,
 		routineClosed:       make(chan struct{}, 1),
 		eventsObj:           events.NewEventsObj(ctx, httpCli),
 		downloadMode:        downloadMode,
-		enableTransactions:  enableTransactions,
+		metrics:             metricsObj,
 		PromMetrics:         promethMetrics,
 	}
 
@@ -95,7 +114,7 @@ func NewBlockAnalyzer(
 	return analyzer, nil
 }
 
-func (s *BlockAnalyzer) Run() {
+func (s *ChainAnalyzer) Run() {
 	defer s.cancel()
 	// Get init time
 	s.initTime = time.Now()
@@ -107,6 +126,9 @@ func (s *BlockAnalyzer) Run() {
 
 	// Metrics per block
 	var wgProcess sync.WaitGroup
+
+	// Validator metrics
+	var wgWorkers sync.WaitGroup
 
 	// Transactions per block
 	var wgTransaction sync.WaitGroup
@@ -127,9 +149,22 @@ func (s *BlockAnalyzer) Run() {
 	}
 	wgProcess.Add(1)
 	go s.runProcessBlock(&wgProcess)
+	wgProcess.Add(1)
+	go s.runProcessState(&wgProcess)
 
 	wgTransaction.Add(1)
 	go s.runProcessTransactions(&wgTransaction)
+
+	for i := 0; i < s.validatorWorkerNum; i++ {
+		// state workers, receiving State and valIdx to measure performance
+		wlog := logrus.WithField(
+			"worker", i,
+		)
+
+		wlog.Tracef("Launching Task Worker")
+		wgWorkers.Add(1)
+		go s.runWorker(wlog, &wgWorkers)
+	}
 
 	s.PromMetrics.Start()
 
@@ -139,9 +174,13 @@ func (s *BlockAnalyzer) Run() {
 	wgProcess.Wait()
 	s.processerFinished = true
 	close(s.blockTaskChan)
+	close(s.epochTaskChan)
 
 	wgTransaction.Wait()
 	close(s.transactionTaskChan)
+
+	wgWorkers.Wait()
+	close(s.valTaskChan)
 
 	s.dbClient.Finish()
 
@@ -151,7 +190,7 @@ func (s *BlockAnalyzer) Run() {
 	s.routineClosed <- struct{}{}
 }
 
-func (s *BlockAnalyzer) Close() {
+func (s *ChainAnalyzer) Close() {
 	log.Info("Sudden closed detected, closing StateAnalyzer")
 	s.stop = true
 	<-s.routineClosed // Wait for services to stop before returning
@@ -166,4 +205,19 @@ type BlockTask struct {
 type TransactionTask struct {
 	Slot         uint64
 	Transactions []*spec.AgnosticTransaction
+}
+
+type EpochTask struct {
+	NextState spec.AgnosticState
+	State     spec.AgnosticState
+	PrevState spec.AgnosticState
+	Finalized bool
+}
+
+type ValTask struct {
+	ValIdxs         []phase0.ValidatorIndex
+	StateMetricsObj metrics.StateMetrics
+	OnlyPrevAtt     bool
+	PoolName        string
+	Finalized       bool
 }
